add code from Sofa and fixs

+++ get_srh_data.py
@@ -2,8 +2,9 @@ import os
 import logging
 import time
 import requests
+from config import delete_fits
 
-def download_file(url, filename, retries=10, timeout=10):
+def download_file(url, filename, retries=10, timeout=45):
     for attempt in range(retries):
         try:
             response = requests.get(url, timeout=timeout)
@@ -19,22 +20,32 @@ def download_file(url, filename, retries=10, timeout=10):
         time.sleep(2 ** attempt)
     return False
 
-def GetSRH(filename, date_str):
+def GetSRH(filename, date_str, max_retries_connect=5):
+
     base_url = 'https://ftp.rao.istp.ac.ru/SRH/corrPlot/'
 
-    try:
-        response = requests.get(base_url, timeout=10)
-        if response.status_code != 200:
-            logging.error(f"Failed to access {base_url}, status code: {response.status_code}")
-            return
-    except requests.RequestException as e:
-        logging.error(f"Error accessing {base_url}: {e}")
-        return
+    for attempt in range(max_retries_connect):
+
+        try:
+            response = requests.get(base_url, timeout=10)
+            if response.status_code != 200:
+                logging.error(f"Failed to access {base_url}, status code: {response.status_code}")
+
+            else:
+                logging.info(f"Status code: {response.status_code}")
+                return
+
+        except requests.RequestException as e:
+            logging.error(f"Error accessing {base_url}: {e}")
+            if attempt < max_retries_connect - 1:
+                time.sleep(2)
+            else:
+                logging.error(f"All {max_retries_connect} attempts check status code failed")
 
     file_url = f'{base_url}/{date_str[0:4]}/{date_str[4:6]}/{filename}'
     if not os.path.exists('data'):
         os.makedirs('data')
-    if os.path.isfile(f'data/{filename}'):
+    if os.path.isfile(f'data/{filename}') and delete_fits:
         os.unlink(f'data/{filename}')
     if not download_file(file_url, f'data/{filename}'):
         logging.error(f"Failed to download {filename} from {file_url}")
